+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-none
+### @dynamic-filtering/core:
+
+#### Changed
+
+-   Fix floating subscriptions to removed filters in the filter manager service - [JHa]
 
 ## [1.0.3] - 02-10-2024